Minor design changes
